move log parsing to a helper fn that's testable
